add Query request

+++ stats-server/src/main/java/com/explore/exploreWithMe/dto/MapperApp.java
@@ -5,6 +5,6 @@ import com.explore.exploreWithMe.model.App;
 
 public class MapperApp {
     public static AppDto mapToAppDto(App app) {
-        return new AppDto(app.getName(), app.getUri(), app.getHit().size());
+        return new AppDto(app.getName(), app.getUri(), (long) app.getHit().size());
     }
 }
